zip/requirements.txt
pillow==8.4.0
tkinter==8.6
